feat: add day 7 part 1 & 2

+++ packages/utils/src/solutions.rs
@@ -1,12 +1,12 @@
 use std::time::Duration;
 
-fn execute_solution(name: &str, closure: Box<dyn FnOnce() -> i32>) -> (&str, i32, Duration) {
+fn execute_solution(name: &str, closure: Box<dyn FnOnce() -> i64>) -> (&str, i64, Duration) {
   let start = std::time::Instant::now();
   let res = closure();
   (name, res, start.elapsed())
 }
 
-pub fn execute_all(closures: Vec<(&str, Box<dyn FnOnce() -> i32>)>) {
+pub fn execute_all(closures: Vec<(&str, Box<dyn FnOnce() -> i64>)>) {
   let start = std::time::Instant::now();
   let mut results = Vec::new();
   for (name, closure) in closures {
